Reorganize folders, bump to go118
