+++ app.testnet.api.yaml
@@ -1,9 +1,8 @@
-runtime: go116
+runtime: go118
 service: api-testnet
 
-main: ./packages/api/cmd
+main: ./packages/api/cmd/web
 
 env_variables:
-  OPENSEA: projects/803405501117/secrets/opensea/versions/1
   PG_CONNSTR: projects/803405501117/secrets/pg-testnet-connstr/versions/1
   NETWORK: testnet
