+++ app.testnet.indexer.yaml
@@ -1,7 +1,7 @@
-runtime: go116
+runtime: go118
 service: indexer-testnet
 
-main: ./packages/api/cmd
+main: ./packages/api/cmd/indexer
 
 instance_class: B2
 
@@ -9,7 +9,6 @@ manual_scaling:
   instances: 1
 
 env_variables:
-  INDEX: True
   OPENSEA: projects/803405501117/secrets/opensea/versions/1
   PG_CONNSTR: projects/803405501117/secrets/pg-testnet-connstr/versions/1
   NETWORK: testnet
