+++ packages/api/game/base.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 
-	"github.com/dopedao/dope-monorepo/packages/api/util"
 	"github.com/gorilla/websocket"
 )
 
@@ -17,7 +16,7 @@ func NewGame() *Game {
 }
 
 func (g *Game) Handle(ctx context.Context, conn *websocket.Conn) error {
-	ctx, log := util.LogFor(ctx)
+	ctx, log := logger.LogFor(ctx)
 	log.Info().Msg("New connection from " + conn.RemoteAddr().String())
 
 	for {
