+++ packages/api/game/player.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"sync"
 
-	"github.com/dopedao/dope-monorepo/packages/api/util"
 	"github.com/google/uuid"
 	"github.com/gorilla/websocket"
 )
@@ -30,7 +29,7 @@ type Player struct {
 }
 
 func (p *Player) readPump(ctx context.Context) {
-	_, log := util.LogFor(ctx)
+	_, log := logger.LogFor(ctx)
 
 	for {
 		var msg BaseMessage
