+++ packages/api/internal/logger/logger.go
@@ -1,9 +1,10 @@
-package util
+package logger
 
 import (
 	"context"
 
 	"github.com/rs/zerolog"
+	"github.com/rs/zerolog/log"
 )
 
 type LogKV func(*zerolog.Context) zerolog.Context
@@ -15,4 +16,11 @@ func LogFor(ctx context.Context, values ...LogKV) (context.Context, zerolog.Logg
 	}
 	lg := sub.Logger()
 	return lg.WithContext(ctx), lg
+}
+
+// Reduce boilerplate code all over the place with this handy function
+func LogFatalOnErr(err error, msg string) {
+	if err != nil {
+		log.Fatal().Err(err).Msgf(msg) //nolint:gocritic
+	}
 }
