+++ packages/api/web/graph/generate.go
