+++ packages/api/web/resources/hustlers.go
@@ -16,7 +16,7 @@ import (
 	"cloud.google.com/go/storage"
 	"github.com/dopedao/dope-monorepo/packages/api/ent"
 	"github.com/dopedao/dope-monorepo/packages/api/ent/hustler"
-	"github.com/dopedao/dope-monorepo/packages/api/util"
+	"github.com/dopedao/dope-monorepo/packages/api/internal/logger"
 	"github.com/gorilla/mux"
 )
 
@@ -35,7 +35,7 @@ type Sprites struct {
 }
 
 func sprite(ctx context.Context, h *ent.Hustler, slot, sex string) string {
-	_, log := util.LogFor(ctx)
+	_, log := logger.LogFor(ctx)
 
 	r := reflect.ValueOf(h.Edges)
 	v := reflect.Indirect(r).FieldByName(slot)
@@ -129,7 +129,7 @@ func sprites(ctx context.Context, id string, client *ent.Client) (*Sprites, erro
 func HustlerSpritesHandler(client *ent.Client) func(http.ResponseWriter, *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
 		ctx := r.Context()
-		ctx, log := util.LogFor(ctx)
+		ctx, log := logger.LogFor(ctx)
 
 		vars := mux.Vars(r)
 		id := vars["id"]
@@ -180,7 +180,7 @@ func readSprite(ctx context.Context, static *storage.BucketHandle, url string) (
 func HustlerSpritesCompositeHandler(client *ent.Client, static *storage.BucketHandle) func(http.ResponseWriter, *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
 		ctx := r.Context()
-		ctx, log := util.LogFor(ctx)
+		ctx, log := logger.LogFor(ctx)
 
 		vars := mux.Vars(r)
 		id := vars["id"]
