+++ packages/api/web/web.go
@@ -1,4 +1,4 @@
-package api
+package web
 
 import (
 	"context"
@@ -15,26 +15,27 @@ import (
 	"github.com/hashicorp/go-retryablehttp"
 	"github.com/rs/cors"
 
+	"github.com/dopedao/dope-monorepo/packages/api"
 	"github.com/dopedao/dope-monorepo/packages/api/authentication"
 	"github.com/dopedao/dope-monorepo/packages/api/engine"
 	"github.com/dopedao/dope-monorepo/packages/api/ent"
-	"github.com/dopedao/dope-monorepo/packages/api/graph"
-	"github.com/dopedao/dope-monorepo/packages/api/middleware"
-	"github.com/dopedao/dope-monorepo/packages/api/resources"
-	"github.com/dopedao/dope-monorepo/packages/api/util"
+	"github.com/dopedao/dope-monorepo/packages/api/internal/logger"
+	"github.com/dopedao/dope-monorepo/packages/api/internal/middleware"
+	"github.com/dopedao/dope-monorepo/packages/api/web/graph"
+	"github.com/dopedao/dope-monorepo/packages/api/web/resources"
 )
 
 // Launch a new HTTP API server to handle web requests
 // for database queries, sprite sheets, authentication, etc.
-func NewHttpServer(ctx context.Context, drv *sql.Driver, static *storage.BucketHandle, network string) (http.Handler, error) {
-	_, log := util.LogFor(ctx)
+func NewServer(ctx context.Context, drv *sql.Driver, static *storage.BucketHandle, network string) (http.Handler, error) {
+	_, log := logger.LogFor(ctx)
 	client := ent.NewClient(ent.Driver(drv))
 
 	// Get Eth client
 	retryableHTTPClient := retryablehttp.NewClient()
 	retryableHTTPClient.Logger = nil
 	// 0 = ethconfig
-	c, err := rpc.DialHTTPWithClient(configs[network][0].(engine.EthConfig).RPC, retryableHTTPClient.StandardClient())
+	c, err := rpc.DialHTTPWithClient(api.AppConfigs[network][0].(engine.EthConfig).RPC, retryableHTTPClient.StandardClient())
 	if err != nil {
 		log.Fatal().Msg("Dialing ethereum rpc.") //nolint:gocritic
 	}
